chore: Remove old lagacy frame
